+++ projects/np-ui-lib/src/lib/np-radio-button/np-radio-group.component.ts
@@ -133,4 +133,8 @@ export class NpRadioGroupComponent implements ControlValueAccessor {
         item.focused = true;
         this.onFocus.emit($event);
     }
+
+    _getInputId(id: string) {
+        return id + "_input";
+    }
 }